Follow up on code review feedback
