+++ apps/react/demo/src/data/index.ts
@@ -1,0 +1,1 @@
+export * from "./examples.js";
