+++ apps/react/demo/src/routes/showcase.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { ShowcaseComponent } from "../ui/index.js";
+import { Showcase as ShowcaseComponent } from "../ui/index.js";
 export const Route = createFileRoute("/showcase")({
   component: Showcase,
 });
