+++ apps/react/demo/src/ui/Example/Context.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { ProviderValue } from "./types.js";
+
+const Context = createContext<ProviderValue | undefined>(undefined);
+
+export default Context;
