+++ apps/react/demo/src/ui/Example/common/Controls/index.ts
@@ -1,0 +1,2 @@
+export { default as Controls } from "./Controls.js";
+export * from "./types.js";
