+++ apps/react/demo/src/ui/Example/common/Controls/types.ts
@@ -1,0 +1,10 @@
+import type { CSSProperties } from "react";
+
+/**
+ * Props for a React component that renders the controls inside a tooltip and allows changing the example configurations.
+ */
+export interface ControlsProps {
+  id?: string;
+  className?: string;
+  style?: CSSProperties;
+}
