+++ apps/react/demo/src/ui/Example/common/Controls/useExampleControls/index.ts
@@ -1,0 +1,1 @@
+export { default as useExampleControls } from "./useExampleControls.js";
