+++ apps/react/demo/src/ui/Example/common/Controls/useExampleControls/types.ts
@@ -1,0 +1,10 @@
+import type { ChangeEvent } from "react";
+
+export type UseExampleControlsResult = {
+  handleFontFamilyChange: (event: ChangeEvent<HTMLSelectElement>) => void;
+  handleFontSizeChange: (event: ChangeEvent<HTMLInputElement>) => void;
+  handleLineHeightChange: (event: ChangeEvent<HTMLInputElement>) => void;
+  handleCopyCss: () => void;
+  handlePrevExample: () => void;
+  handleNextExample: () => void;
+};
