+++ apps/react/demo/src/ui/Example/common/Renderer/index.ts
@@ -1,0 +1,2 @@
+export { default as Renderer } from "./Renderer.js";
+export * from "./types.js";
