+++ apps/react/demo/src/ui/Example/common/examples/index.ts
@@ -1,4 +1,3 @@
 // button is a temporary example just so we have multiple examples to show in the showcase
 export { default as Button } from "./Button.js";
 export * from "./TypographicSpecimen/index.js";
-export * from "./examples.js";
