+++ apps/react/demo/src/ui/Example/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./useConfig/index.js";
