+++ apps/react/demo/src/ui/Example/hooks/useConfig/index.ts
@@ -1,0 +1,1 @@
+export { default as useConfig } from "./useConfig.js";
