+++ apps/react/demo/src/ui/Example/utils/index.ts
@@ -1,0 +1,1 @@
+export { default as generateCssVariables } from "./generateCssVariables.js";
