--- apps/react/demo/src/ui/Showcase/common/ExampleControls/index.ts
@@ -1,3 +1,0 @@
-export { default as ExampleControls } from "./ExampleControls.js";
-export { ConfigProvider, useConfig } from "./Context.js";
-export * from "./types.js";
