--- apps/react/demo/src/ui/Showcase/common/ExampleControls/styles.css
@@ -1,10 +1,0 @@
-.ds.example-controls {
-  background-color: black;
-  width: 100%;
-  padding: var(--spacing-horizontal-small) var(--spacing-vertical-small);
-}
-
-.ds.example-controls__config-button {
-  /* Configure button has collapsed margins - add margin bottom to make it consistent with other buttons */
-  margin-bottom: var(--spacing-vertical-small);
-}
