--- apps/react/demo/src/ui/Showcase/common/ExampleRenderer/index.ts
@@ -1,2 +1,0 @@
-export { default as ExampleRenderer } from "./ExampleRenderer.js";
-export * from "./types.js";
