--- apps/react/demo/src/ui/Showcase/common/index.ts
@@ -1,3 +1,0 @@
-export * from "./ExampleControls/index.js";
-export * from "./ExampleRenderer/index.js";
-export * from "./examples/index.js";
