+++ apps/react/demo/src/ui/Showcase/index.ts
@@ -1,2 +1,1 @@
-export { default as ShowcaseComponent } from "./Showcase.js";
-export * from "./common/index.js";
+export { default as Showcase } from "./Showcase.js";
